feat: add product modal form
