+++ src/contexts/AuthProvider.tsx
@@ -86,7 +86,7 @@ export function AuthProvider({ children }: { children: ReactNode }) {
       isMounted = false
       subscription.unsubscribe()
     }
-  }, [])
+  }, [logger])
 
   const signIn = async (email: string, password: string) => {
     try {
